+++ lte/alt1250/usock_handlers/alt1250_closehdlr.c
@@ -78,7 +78,7 @@ int postproc_sockcommon(FAR struct alt1250_s *dev,
                         FAR struct alt_container_s *reply,
                         FAR struct usock_s *usock,
                         FAR int32_t *usock_result,
-                        FAR uint64_t *usock_xid,
+                        FAR uint32_t *usock_xid,
                         FAR struct usock_ackinfo_s *ackinfo,
                         unsigned long arg)
 {
@@ -123,7 +123,7 @@ int postproc_sockcommon(FAR struct alt1250_s *dev,
 int usockreq_close(FAR struct alt1250_s *dev,
                    FAR struct usrsock_request_buff_s *req,
                    FAR int32_t *usock_result,
-                   FAR uint64_t *usock_xid,
+                   FAR uint32_t *usock_xid,
                    FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_close_s *request =
