+++ lte/alt1250/usock_handlers/alt1250_connecthdlr.c
@@ -52,7 +52,7 @@ static int postproc_getsockopt(FAR struct alt1250_s *dev,
                                FAR struct alt_container_s *reply,
                                FAR struct usock_s *usock,
                                FAR int32_t *usock_result,
-                               FAR uint64_t *usock_xid,
+                               FAR uint32_t *usock_xid,
                                FAR struct usock_ackinfo_s *ackinfo,
                                unsigned long arg)
 {
@@ -109,7 +109,7 @@ static int postproc_connect(FAR struct alt1250_s *dev,
                             FAR struct alt_container_s *reply,
                             FAR struct usock_s *usock,
                             FAR int32_t *usock_result,
-                            FAR uint64_t *usock_xid,
+                            FAR uint32_t *usock_xid,
                             FAR struct usock_ackinfo_s *ackinfo,
                             unsigned long arg)
 {
@@ -225,7 +225,7 @@ int nextstep_connect(FAR struct alt1250_s *dev,
                      FAR struct alt_container_s *reply,
                      FAR struct usock_s *usock,
                      FAR int32_t *usock_result,
-                     FAR uint64_t *usock_xid,
+                     FAR uint32_t *usock_xid,
                      FAR struct usock_ackinfo_s *ackinfo,
                      unsigned long arg)
 {
@@ -251,7 +251,7 @@ int nextstep_connect(FAR struct alt1250_s *dev,
 int usockreq_connect(FAR struct alt1250_s *dev,
                      FAR struct usrsock_request_buff_s *req,
                      FAR int32_t *usock_result,
-                     FAR uint64_t *usock_xid,
+                     FAR uint32_t *usock_xid,
                      FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_connect_s *request = &req->request.conn_req;
