+++ lte/alt1250/usock_handlers/alt1250_getpeernamehdlr.c
@@ -42,7 +42,7 @@
 int usockreq_getpeername(FAR struct alt1250_s *dev,
                          FAR struct usrsock_request_buff_s *req,
                          FAR int32_t *usock_result,
-                         FAR uint64_t *usock_xid,
+                         FAR uint32_t *usock_xid,
                          FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_getpeername_s *request =
