+++ lte/alt1250/usock_handlers/alt1250_sendtohdlr.c
@@ -85,7 +85,7 @@ static int send_sendto_command(FAR struct alt1250_s *dev,
 int usockreq_sendto(FAR struct alt1250_s *dev,
                     FAR struct usrsock_request_buff_s *req,
                     FAR int32_t *usock_result,
-                    FAR uint64_t *usock_xid,
+                    FAR uint32_t *usock_xid,
                     FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_sendto_s *request = &req->request.send_req;
